+++ packages/synapse-interface/components/_Transaction/helpers/getExplorerAddressLink.ts
@@ -2,16 +2,14 @@ import { Chain } from '@/utils/types'
 import { CHAINS_BY_ID } from '@/constants/chains'
 
 export const getExplorerAddressLink = (chainId: number, address: string) => {
-  const chain: Chain = CHAINS_BY_ID[chainId]
-
-  if (chain && address) {
-    const addressExplorerUrl = `${chain.explorerUrl}/address/${address}`
-    const explorerName = chain.explorerName
-
-    return [addressExplorerUrl, explorerName]
+  if (!chainId || !address) {
+    return [null, null]
   }
 
-  console.error('getExplorerAddressLink: Chain or Address missing')
+  const chain: Chain = CHAINS_BY_ID[chainId]
 
-  return [null, null]
+  const addressExplorerUrl = `${chain.explorerUrl}/address/${address}`
+  const explorerName = chain.explorerName
+
+  return [addressExplorerUrl, explorerName]
 }