+++ packages/synapse-interface/store/reducer.ts
@@ -15,6 +15,7 @@ import poolUserData from '@/slices/poolUserDataSlice'
 import poolWithdraw from '@/slices/poolWithdrawSlice'
 import priceData from '@/slices/priceDataSlice'
 import swapDisplay from '@/slices/swapDisplaySlice'
+import feeAndRebate from '@/slices/feeAndRebateSlice'
 import { api } from '@/slices/api/slice'
 import { RootActions } from '@/slices/application/actions'
 
@@ -27,7 +28,7 @@ const persistedReducers = {
 export const storageKey: string = 'synapse-interface'
 
 export const persistConfig: PersistConfig<AppState> = {
-  version: 1, // upgrade to reset cache when updated data structures throw errors
+  version: 2, // upgrade to reset cache when updated data structures throw errors
   key: storageKey,
   storage,
   whitelist: Object.keys(persistedReducers),
@@ -44,6 +45,7 @@ export const appReducer = combineReducers({
   poolWithdraw,
   priceData,
   swapDisplay,
+  feeAndRebate,
   [api.reducerPath]: api.reducer,
   ...persistedReducers,
 })